+++ docs/release_notes.rst
@@ -65,6 +65,25 @@ What's New?
 *   Migrating :class:`.DataZip` functionality to :class:`etoolbox.datazip.DataZip`.
 *   Updates to constants to allow Nuclear and Conventional Hydroelectric to be properly
     displayed in plots.
+*   Updates to ``re_plant_specs``, its validation, and
+    :meth:`.DispatchModel.re_and_net_load` for a new column, ``interconnect_mw``, that
+    allows interconnection capacity for a renewable facility to independent of its
+    capacity. By default, this is the same as ``capacity_mw`` but can be reduced to
+    reflect facility-specific transmission / interconnection constraints. If the
+    facility has storage, storage can be charged by the constrained excess.
+
+
+Bug Fixes
+^^^^^^^^^
+*   Fixed a bug where storage metrics in :meth:`.DispatchModel.system_level_summary`
+    were :class:`numpy.nan` because selecting of data from ``storage_specs`` returned
+    a :class:`pandas.Series` rather than a :class:`int` or :class:`float`. Further, in
+    cases of division be zero in these calculations, the result is now 0 rather than
+    :class:`numpy.nan`. Tests now make sure that no new :class:`numpy.nan` show up.
+*   Fixed a bug in :meth:`.DispatchModel.dispatchable_summary` where ``pct_replaced``
+    would be :class:`numpy.nan` because of division be zero in these calculations, the
+    result is now 0 rather than :class:`numpy.nan`. Tests now make sure that no new
+    :class:`numpy.nan` show up.
 
 Known Issues
 ^^^^^^^^^^^^
@@ -76,7 +95,8 @@ Known Issues
 *   It is possible that output from DC-coupled RE+Storage facilities during some hours
     will exceed the system's inverter capacity because when we discharge these storage
     facilities, we do not know how much 'room' there is in the inverter because we do
-    not know the RE-side's output.
+    not know the RE-side's output. This issue is now in some sense compounded when
+    ``interconnect_mw`` is less than ``capacity_mw``.
 *   :class:`.DataZip` are effectively immutable once they are created so the ``a`` mode
     is not allowed and the ``w`` mode is not allowed on existing files. This is because
     it is not possible to overwrite or remove a file already in a